pyproject.toml
[project]
name = "code-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-search-results>=2.4.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
]
